First pass on launcher, host dll, and plugin dlls
Launcher: Starts RFGR in a suspended state and injects the host DLL into it. Eventually will have a gui.
Host DLL: Loads plugins and handles reloading them when necessary
Plugin DLLs: Most code will be in these. No hooking or patches yet, but it'll be in these. Hopefully will be able to hot reload DLLs with patches safely but that might prove to be difficult.