Organize code a bit
Need to think about how this is organized some more. This works but it seems like it'll become a pain as the project grows. The new code still needs to be properly commented but dear imgui + overlays are in a good enough working that I'll push this and merge this branch into master now.