Simplify plugin interfaces
Instead of exporting individual functions they instead define a single struct/class with a bunch of function pointers and export that. Any plugin that wants to use that interface will need it's header. This is an improvement since it avoids having a ton of boilerplate to define function pointer types and names and manually exporting/importing each function. Just have to define/export/import a single struct of function pointers.