Fix bug with global state initialization
Some global values would fail to be initialized on the first call of RSL2_PluginInit() because none of the RFG init code had run yet. This commit tries to fix that by initializing the globals again after one of RFGs main initialization functions is done running. So far this seems to have fixed the issue.