Initial setup
Got the basics compiling and running correctly. Next up I need to bind some of the game to python and evaluate if it's capable enough for the script loader.

Pythons syntax is nice and it has a large standard library so I think it's a good choice. It should also be easy to bind to c++ due to pybind. 

One concern I have if it's performant enough. A common thing I expect mods to do is to iterate all objects of a certain type each frame and perform some operation on them. It should be possible to have many scripts doing this without tanking the framerate. So a decent target is that a script should be able to iterate all objects of a certain type in much less than 16.6ms (the frametime at 60FPS). 

If it can't do that some speedup might be possible by exposing common operations as native functions to python. There's also JIT'ed variants of python like PyPy. Then finally there's LuaJIT if python doesn't cut it. 

One problem I don't have an answer for yet is distributing python. I'd like it if neither developers nor users of RSL2 had to manually install python. It would make installation easier and avoid issues with people installing incompatible python versions. I'm currently building off my systems python install but I'll test out packaging a build of python with RSL2 soon.