Working test of a custom game state
Test of adding a custom state to the game and pushing it onto the state stack. This code replaces the main menu state the first time it's loaded so that when the game is started it goes to an imgui window instead of the main menu. 

No real use for it now, just wanted to see if it would work. Might be useful for adding custom modes or tools in the game later on.