Fix player* being out of date in certain cases
The global player* changes when the player dies or when a mission or activity is cancelled. RSL wasn't updating it's copy of the pointer when it changes. This commit fixes that. We could use a player** to avoid the need for updating it, but I'd rather not have to use pointer pointers.