Add DevTools plugin
This plugin will hold built in GUIs and tools, such as the dev console and guis for debugging mods and scripts. Inclusion of UIs will be more selective than RSL1. The trainer UIs it had will be separate mods for RSL2. 

This will require binding dear imgui to lua. We can start small and gradually add the API since it's fairly large.