Increase default size of many memory pools and string pools
Mods were hitting the limits of some of these pools. The game has a lot of fixed size pools with just enough room to hold the vanilla game. This is a problem when mods start to add things to the game. There's still a few pools to increase the size of and some harder things to patch like static arrays and generic types with the size as a template argument.