Overlay code revisions
Changes made in to prepare for DevTools.dll, which will hold dev/modder tools and be an optional install. Consolidated the multiple imgui callbacks into one GuiBase class which can be inherited and overriden to add overlay guis and widgets. Originally I went with C style interface functions so plugins could be written in languages other than C++. Now I'm not so concerned with that and would rather have a proper C++ class so the code is easier to maintain.

Other languages can still interop with this. It'll just be slightly more difficult since they have to recreate the vtable on their side. Not a big concern for now with how few modders RFG has.